+++ development/application/api/http_root/__init__.py
@@ -2,6 +2,7 @@
 # coding: utf-8
 
 # Namespace Imports
+import cronjobs
 from jsonrpc import jsonrpc