+++ development/application/common/named_values.py
@@ -3,13 +3,20 @@
 
 from model.named_value import NamedValue
 
+_global_keys = {}
+
 
 def get_value(name):
     """
     Returns the value with the given name.
     """
 
-    return NamedValue.query(NamedValue.name == name).get()
+    try:
+        return _global_keys[name].get()
+    except KeyError:
+        named_value = NamedValue.query(NamedValue.name == name).get()
+        _global_keys[name] = named_value.key
+        return named_value
 
 
 def set_value(name, value):
@@ -19,10 +26,13 @@ def set_value(name, value):
 
     # Get existing named value
     named_value = get_value(name)
+
     if not named_value:
         # Create new named value
         named_value = NamedValue()
         named_value.name = name
+
+    # Edit value
     named_value.value = value
 
     # Save