+++ development/history.rst
@@ -7,6 +7,16 @@ ToDo: Security: Authorization: *common.authorization*
 
 
 =============
+Version 0.2.3
+=============
+
+2016-05-23
+
+
+
+
+
+=============
 Version 0.2.2
 =============